\name{qtlnet.prior-internal}
\alias{check.qtlnet}
\alias{find.index.parent}
\alias{find.parent.score}
\alias{forbidden.additions}
\alias{geno.effect}
\alias{index.parents}
\alias{scan.pheno}
\alias{check.additions}
\alias{check.cycle}
\alias{check.downstream}
\alias{check.reversions}
\alias{check.reversions.edge}
\alias{check.upstream}
\alias{update.node}
\alias{drop.edge}
\alias{sample.exclude.down}
\alias{propose.add}
\alias{rev.edge}
\alias{reverse.proposal}
\alias{bic.qtlnet.pheno}
\alias{upstream.additions}
\alias{check.upstream.reversions}
\alias{propose.new.structure}
\alias{propose.new.node.edge}
\alias{nbhd.size}
\alias{upstream.additions}
\alias{check.upstream.reversions}
\alias{calc.pheno.bic}
\title{Internal qtlnet.prior functions}
\details{
	These are not to be called by the user.
}

\description{
	Internal use
}
\keyword{ internal }

