\name{sim.example1.cross}
\alias{sim.example1.cross}
\title{
  Generate an example of cross of 500 individuals with 5 phenotypes
}
\description{
	It generates an F2 cross object consisting of genetic information
	and 5 phenotypes for 500 individuals. See the details for the detailed information.  
}
\usage{
sim.example1.cross(seed = 1234)
%geno.effect(genotype, additive, dominance)
}

\arguments{
  \item{seed}{
  Random seed number}
}
\details{
	\code{sim.example1.cross} generates an F2 cross object consisting of genetic information
	and 5 phenotypes for 500 individuals.  
	
	The underlying causal network between QTL and phenotypes is as following:
	Q1 -> Y1, (Q2, Y1) -> Y2, Y1 -> Y3, (Q4, Y1, Y3) -> Y4, (Q5, Y2, Y3, Y4) -> Y5.
    Q_t is a QTL in the middle of chromosome_t.  
    
    First, the genetic map is generated for 5 autosomal chromosomes of 100 cM with 10 unequally spaced markers in each chromosome.  
    
	Second, genetic additive effects, dominance effects, and partial regression coefficients are randomly generated. 
	Genetic additive effects are sampled from U[0, 0.5], genetic dominance effects are sampled from U[0, 0.25], and partial regression
	coefficients are sampled from U[-0.5, 0.5].  
	
	Third, phenotypes are generated by adding a Gaussian noise of variance 1 to the calculated phenotypic value (immediate QTL effect + immediate upstream phenotypic effect). 	
}
\value{
	An object of class \code{cross}, which is a list of three components:
	\item{geno}{A list of genetic information for each chromosome}
	\item{pheno}{A matrix of QTL genotypes}
	\item{qtlgeno}{A data.frame of phenotypes}
}


\seealso{
	\code{\link{sim.cross}},
	\code{\link{sim.map}},
	\code{\link{read.cross}}
}
\examples{
sim.example1.cross(356)
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
